tools: Add os.py for os-related utilities
