+++ devito/tools/__init__.py
@@ -4,5 +4,6 @@ from devito.tools.abc import *  # noqa
 from devito.tools.algorithms import *  # noqa
 from devito.tools.data_structures import *  # noqa
 from devito.tools.memoization import *  # noqa
+from devito.tools.os import *  # noqa
 from devito.tools.validators import *  # noqa
 from devito.tools.visitors import *  # noqa
